+++ tests/HeadTest.php
@@ -9,3 +9,11 @@ test( 'head', function() {
 	expect( $response['headers']['response_code'] )->toBe( 200 );
 	expect( $response['headers']['content-type'] )->toBe( 'application/json' );
 } );
+
+test( 'head: static', function() {
+	$response = HTTP::head( url: 'http://127.0.0.1:7878/?method=head' );
+
+	expect( $response['error'] )->toBe( false );
+	expect( $response['headers']['response_code'] )->toBe( 200 );
+	expect( $response['headers']['content-type'] )->toBe( 'application/json' );
+} );