+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## ????
 - Major changes to the API, reducing the amount of code and making the API more consistent across the static and object methods
+- Static tests have been merged into a single set per verb
 - Simplify body encoding checks
 - Makefile test failure cleanup improved